refactor(state): remove unused imports
